corrections

+++ delivery/src/main/java/com/solvd/DeliveryService/model/interface1/IOperate.java
@@ -2,5 +2,5 @@ package com.solvd.DeliveryService.model.interface1;
 
 @FunctionalInterface
 public interface IOperate {
-    int operate(int Distance, int speed);
+    int operate(double Distance, int speed);
 }
